Fixed background
I found some bugs in the background CSS while cleaning it. So I have
fixed it with adding a div into the body and placing the background on
that. I know it isn't a clean fix, but I do believe it is the cleanest
possible fix. I've tried over 20 possible way to fix it without use of a
helper html object, but none of them were successful... (CSS sucks)